Yank $createdAt out of if/else statement.
Add another condition to the if/else statement. I queried the Magento database because I was curious how an empty string could get put in store_id. It still doesn't actually make sense because it looks like the store_id column in the store table is supposed to be non-null, but that empty value is coming from somewhere. Either way, there's an admin store in there that doesn't seem to appear in the actual magento interface that's `store_id = 0`. I'm guessing it shouldn't be used for things in production, but it could theoretically still be passed to us, and `empty()` will still evaluate `"0"` as an empty value.